extract uuid input
