+++ src/components/Auth/ValidateEmail.jsx
@@ -1,5 +1,4 @@
-/* eslint react/no-array-index-key: 0 */
-import React, { Component, Fragment } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
 import { withRouter } from 'react-router';
@@ -8,32 +7,13 @@ import Grid from 'material-ui/Grid';
 import Button from 'material-ui/Button';
 import FormGroup from 'material-ui/Form/FormGroup';
 import { red } from 'material-ui/colors';
-import TextField from 'material-ui/TextField';
 import { withStyles } from 'material-ui/styles';
-import { compose, withStateHandlers /* withProps */ } from 'recompose';
+import { compose, withStateHandlers } from 'recompose';
 
-// import fetch from 'unfetch';
 import { graphql } from 'react-apollo';
 import gql from 'graphql-tag';
 import validateEmail from '../../../graphql/mutations/validateEmail';
-
-const uuidSecretAttrs = [
-	{
-		length: 8
-	},
-	{
-		length: 4
-	},
-	{
-		length: 4
-	},
-	{
-		length: 4
-	},
-	{
-		length: 12
-	}
-];
+import { UUIDInput, uuidInputEnhancer } from './UUIDInput';
 
 const styles = theme => ({
 	root: {
@@ -59,141 +39,61 @@ const styles = theme => ({
 	}
 });
 
-class ValidateEmail extends Component {
-	static propTypes = {
-		classes: PropTypes.object.isRequired,
-		validationSecret: PropTypes.arrayOf(PropTypes.string).isRequired,
-		validationError: PropTypes.arrayOf(PropTypes.bool).isRequired,
-		uuidAttrs: PropTypes.arrayOf(PropTypes.object),
-		submit: PropTypes.func.isRequired,
-		disabled: PropTypes.bool.isRequired,
-		attemptedValidation: PropTypes.bool.isRequired,
-		validationSuccess: PropTypes.bool.isRequired,
-		onFieldChange: PropTypes.func.isRequired,
-		onFocusChange: PropTypes.func.isRequired,
-		focus: PropTypes.number.isRequired
-	};
-	static defaultProps = {
-		uuidAttrs: uuidSecretAttrs
-	};
-	constructor(...args) {
-		super(...args);
-		this._inputs = {};
-	}
-	componentDidUpdate(prevProps) {
-		const { focus: prevFocus } = prevProps;
-		const { focus: newFocus, onFocusChange } = this.props;
-		if (prevFocus !== newFocus) {
-			try {
-				this._inputs[newFocus].focus();
-				onFocusChange(newFocus);
-			} catch (err) {
-				// nothing
-			}
-		}
-	}
-	render() {
-		const {
-			classes,
-			validationSecret,
-			submit,
-			disabled,
-			attemptedValidation,
-			validationSuccess,
-			onFieldChange,
-			validationError,
-			uuidAttrs
-		} = this.props;
-		return (
+const ValidateEmail = ({
+	classes,
+	submit,
+	uuidIncomplete,
+	attemptedValidation,
+	validationSuccess,
+	validationSecret,
+	onFieldChange,
+	validationError,
+	focus
+}) => (
   <Grid className={classes.root} container>
     <Grid container item xs={12} justify="center">
       <FormGroup className={classes.form}>
         {attemptedValidation && validationSuccess ? (
           <h6>
-								Email Verified. If you are not redirected to login page,{' '}
+						Email Verified. If you are not redirected to login page,{' '}
             <Link to="/login">click here</Link>
           </h6>
-						) : (
+				) : (
   <h3>Enter your validation key we send to your email</h3>
-						)}
-        <FormGroup className={classes.sercet}>
-          {uuidAttrs.map((attr, index) => (
-            <Fragment>
-              <TextField
-                key={index}
-                autoFocus={index === 0}
-                value={validationSecret[index]}
-                error={validationError[index]}
-                inputProps={{
-											'data-section': index,
-											'data-length': attr.length
-										}}
-                onChange={onFieldChange}
-                inputRef={node => {
-											this._inputs = this._inputs || {};
-											this._inputs[index] = node;
-										}}
-              />
-              {index !== 4 && '-'}
-            </Fragment>
-							))}
-        </FormGroup>
+				)}
+        <UUIDInput
+          focus={focus}
+          className={classes.sercet}
+          validationSecret={validationSecret}
+          validationError={validationError}
+          onFieldChange={onFieldChange}
+        />
         <Button
           variant="raised"
           label="Create Clipboard"
           color="primary"
           onClick={submit}
-          disabled={disabled}
-          ref={node => {
-								this._inputs[5] = node;
-							}}
+          disabled={uuidIncomplete}
         >
-							validate my email
+					validate my email
         </Button>
         {attemptedValidation && !validationSuccess && <h6>Check entry</h6>}
       </FormGroup>
     </Grid>
   </Grid>
-		);
-	}
-}
+);
 
-const onFieldChange = ({ validationSecret, validationError }) => event => {
-	const prevSecret = validationSecret;
-	const { dataset: data = {}, value = '' } = event.target;
-	const section = parseInt(data.section, 10);
-	let newFocus = section;
-	// handle user input something
-	const newSecret = [...prevSecret];
-	newSecret[section] = value;
-	// handle error (not number or lowercase)
-	const length = parseInt(data.length, 10);
-	const alphaNumOnly = /^[0-9a-z]*$/.test(value);
-	const newValidationError = [...validationError];
-	newValidationError[section] = value === '' ? false : length < value.length || !alphaNumOnly;
-	// try to jump focus
-	const prevSectionValue = prevSecret[section];
-	const newSectionValue = newSecret[section];
-	if (prevSectionValue !== '' && newSectionValue === '' && section !== 0) {
-		newFocus = section - 1;
-	} else if (newSectionValue.length === length && !newValidationError[section]) {
-		newFocus = section + 1;
-	}
-	// see if user is ready
-	const noError = newValidationError.every(acc => acc === false);
-	const sercetFilled = newSecret.every(
-		(secret, i) => secret.length === uuidSecretAttrs[i].length
-	);
-	const disabled = !noError || !sercetFilled;
-	return {
-		validationSecret: newSecret,
-		validationError: newValidationError,
-		focus: newFocus,
-		disabled
-	};
+ValidateEmail.propTypes = {
+	classes: PropTypes.object.isRequired,
+	validationSecret: PropTypes.arrayOf(PropTypes.string).isRequired,
+	validationError: PropTypes.arrayOf(PropTypes.bool).isRequired,
+	onFieldChange: PropTypes.func.isRequired,
+	submit: PropTypes.func.isRequired,
+	uuidIncomplete: PropTypes.bool.isRequired,
+	attemptedValidation: PropTypes.bool.isRequired,
+	validationSuccess: PropTypes.bool.isRequired,
+	focus: PropTypes.number.isRequired
 };
-
-const onFocusChange = () => newFocus => ({ focus: newFocus });
 
 const onValidationError = () => () => ({ attemptedValidation: true, validationSuccess: false });
 const onValidationSuccess = () => () => ({ attemptedValidation: true, validationSuccess: true });
@@ -230,30 +130,21 @@ const withValidateEmail = graphql(
 
 const recomposeEnhancer = compose(
 	withStateHandlers(
-		({
-			validationSecret = new Array(5).fill(''),
-			validationError = new Array(5).fill('').map(() => false),
-			disabled = true,
-			focus = -1,
-			attemptedValidation = false,
-			validationSuccess = false
-		}) => ({
-			validationSecret,
-			validationError,
-			disabled,
-			focus,
+		({ attemptedValidation = false, validationSuccess = false }) => ({
 			attemptedValidation,
 			validationSuccess
 		}),
 		{
-			onFieldChange,
-			onFocusChange,
 			onValidationError,
 			onValidationSuccess
 		}
 	)
 );
 
-export default compose(withStyles(styles), withRouter, recomposeEnhancer, withValidateEmail)(
-	ValidateEmail
-);
+export default compose(
+	withStyles(styles),
+	withRouter,
+	uuidInputEnhancer,
+	recomposeEnhancer,
+	withValidateEmail
+)(ValidateEmail);
